{"schema":"decodex/commit/1","summary":"Modularize autonomy contracts","authority":"manual"}
