+++ apps/decodex/src/autonomy_signal/validation.rs
@@ -1,0 +1,35 @@
+//! Shared autonomy signal validation helpers.
+
+use crate::prelude::{Result, eyre};
+
+pub(super) fn validate_required(name: &str, value: &str) -> Result<()> {
+	if value.trim().is_empty() {
+		eyre::bail!("{name} must not be empty.");
+	}
+
+	Ok(())
+}
+
+pub(super) fn validate_optional_required(name: &str, value: Option<&str>) -> Result<()> {
+	if let Some(value) = value {
+		validate_required(name, value)?;
+	}
+
+	Ok(())
+}
+
+pub(super) fn validate_string_list(name: &str, values: &[String]) -> Result<()> {
+	for value in values {
+		validate_required(name, value)?;
+	}
+
+	Ok(())
+}
+
+pub(super) fn validate_nonempty_list(name: &str, values: &[String]) -> Result<()> {
+	if values.is_empty() {
+		eyre::bail!("{name} must not be empty.");
+	}
+
+	validate_string_list(name, values)
+}
